Polish the VR integration (#10)
* Fix per-eye culling

* Swapping hands when dual elites require it

* Rifle zooming by motion

* Arm length adjusting added

* OpenXR fixes

* Rifle zooming better

* Motion controls fixes

* Move the VR code into its own file
