+++ engine/common/common.h
@@ -568,14 +568,16 @@ void Host_Credits( void );
 void Host_ExitInMain( void );
 
 void Host_VRInit( void );
-void Host_VRInput( void );
-void Host_VRButtonMap( int button, int currentButtons, int lastButtons, const char* name, bool alt );
+bool Host_VRInitFrame( void );
+void Host_VRClientFrame( void );
+void Host_VRInputFrame( void );
+void Host_VRButtonMap( unsigned int button, int currentButtons, int lastButtons, const char* name, bool alt );
 void Host_VRButtonMapping( bool swapped, int lbuttons, int rbuttons );
 bool Host_VRConfig();
 void Host_VRCursor( bool cursorActive, float x, float y, vec2_t cursor );
 void Host_VRCustomCommand( char* action );
 bool Host_VRMenuInput( bool cursorActive, bool gameMode, bool swapped, int lbuttons, int rbuttons, vec2_t cursor );
-void Host_VRMotionControls( vec3_t hmdAngles, vec3_t handPosition, vec3_t hmdPosition, vec3_t weaponPosition);
+void Host_VRMotionControls( bool zoomed, bool superzoomed, vec3_t hmdAngles, vec3_t handPosition, vec3_t hmdPosition, vec3_t weaponPosition );
 void Host_VRMovementEntity( bool zoomed, vec3_t handPosition, vec3_t hmdAngles, vec3_t hmdPosition, vec3_t weaponPosition );
 void Host_VRMovementPlayer( vec3_t hmdAngles, vec3_t hmdPosition, vec3_t weaponAngles, float thumbstickX, float thumbstickY );
 void Host_VRRotations( bool zoomed, vec3_t handAngles, vec3_t hmdAngles, vec3_t hmdPosition, vec3_t weaponAngles, float thumbstickX, float thumbstickY );
